pyproject.toml
[project]
name = "gerar-classes-pydantic"
version = "0.1.0"
description = ""
authors = [
    {name = "ThiagoHolandaRamalho",email = "thiago.holandarm@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
